# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  firebaseUid: string;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  displayName: string | null;
  /** @nullable */
  photoUrl: string | null;
  role: string;
  plan: string;
  disabled: boolean;
  createdAt: string;
  lastLoginAt: string;
}