# lib/api-zod/src/generated/types/apiKeyCreated.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyCreated {
  id: number;
  name: string;
  keyPrefix: string;
  active: boolean;
  key: string;
}