# lib/api-zod/src/generated/types/apiKeyInfo.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyInfo {
  id: number;
  name: string;
  keyPrefix: string;
  active: boolean;
  createdAt: string;
  /** @nullable */
  lastUsedAt: string | null;
}