# lib/api-zod/src/generated/types/premiumPackage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PremiumPackage {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  priceCents: number;
  currency: string;
  periodDays: number;
  active: boolean;
}