# lib/api-zod/src/generated/types/premiumPackageInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PremiumPackageInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  priceCents: number;
  currency: string;
  /** @minimum 1 */
  periodDays: number;
  active: boolean;
}