# lib/api-zod/src/generated/types/usageStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UsageStatsTopToolsItem } from './usageStatsTopToolsItem';

export interface UsageStats {
  totalUsers: number;
  totalEvents: number;
  eventsLast7Days: number;
  topTools: UsageStatsTopToolsItem[];
}